django_api/requirements.txt
Django>=3.0,<6.0
djangorestframework>=3.12.0
django-cors-headers==4.4.0